requirements.txt
python-dotenv
openai
openai-whisper @ git+https://github.com/openai/whisper.git
fuzzywuzzy
python-Levenshtein
transformers
pyaudio
keyboard
pygame
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2